Don't count comments on own PRs
The count of reviews and comments was including an author's comments on
their own PRs. These occur when they respond to comments. A response can
be done as a single published review, or comment-by-comment. In the
latter case, each comment is counted as a separate review. This can
massively inflate the numbers for peoples responses to reviews on their
PRs and is unfavourable to those with no PRs or few comments on their
PRs.

Update the tests to add the case of self-comments on a PR to ensure they
are not counted.